+++ tests/toisto/ui/test_text.py
@@ -2,7 +2,6 @@
 
 from unittest import TestCase
 
-from toisto.model.language import Language
 from toisto.model.language.label import Label
 from toisto.model.quiz.quiz_factory import create_quizzes
 from toisto.persistence.spelling_alternatives import load_spelling_alternatives
@@ -19,14 +18,14 @@ class FeedbackTestCase(ToistoTestCase):
     def setUp(self) -> None:
         """Override to set up test fixtures."""
         super().setUp()
-        load_spelling_alternatives(Language(self.fi), Language(self.nl))
-        self.guess = Label(Language(self.fi), "terve")
+        load_spelling_alternatives(self.fi_nl)
+        self.guess = Label(self.fi_nl.target, "terve")
 
     def test_correct_first_time(self):
         """Test that the correct feedback is given when the user guesses correctly."""
         concept = self.create_concept("hi", dict(nl="hoi", fi="terve"))
-        quiz = create_quizzes(self.nl, self.fi, concept).by_quiz_type("read").pop()
-        feedback_text = feedback_correct(self.guess, quiz, self.nl)
+        quiz = create_quizzes(self.nl_fi, concept).by_quiz_type("read").pop()
+        feedback_text = feedback_correct(self.guess, quiz, self.nl_fi)
         self.assertEqual(CORRECT, feedback_text)
 
     def test_show_colloquial_language(self):
@@ -38,32 +37,34 @@ class FeedbackTestCase(ToistoTestCase):
         expected_feedback_correct = CORRECT + colloquial + meaning
         expected_feedback_incorrect = INCORRECT + answer + colloquial + meaning
         expected_feedback_on_skip = f'The correct answer is "{linkify("kiitos")}".\n' + colloquial + meaning
-        for quiz in create_quizzes(self.fi, self.nl, concept).by_quiz_type("dictate"):
+        for quiz in create_quizzes(self.fi_nl, concept).by_quiz_type("dictate"):
             if quiz.question.is_colloquial:
-                self.assertEqual(expected_feedback_correct, feedback_correct(Label(self.fi, "kiitos"), quiz, self.fi))
+                self.assertEqual(
+                    expected_feedback_correct, feedback_correct(Label(self.fi, "kiitos"), quiz, self.fi_nl)
+                )
                 self.assertEqual(expected_feedback_incorrect, feedback_incorrect(Label(self.fi, "hei"), quiz))
                 self.assertEqual(expected_feedback_on_skip, feedback_incorrect(Label(self.fi, "?"), quiz))
 
     def test_show_alternative_answer(self):
         """Test that alternative answers are shown."""
         concept = self.create_concept("hi", dict(nl="hoi", fi=["terve", "hei"]))
-        quiz = create_quizzes(self.nl, self.fi, concept).by_quiz_type("read").pop()
+        quiz = create_quizzes(self.nl_fi, concept).by_quiz_type("read").pop()
         expected_other_answer = linkify(quiz.other_answers(self.guess)[0])
         expected_text = f'{CORRECT}[{SECONDARY}]Another correct answer is "{expected_other_answer}".[/{SECONDARY}]\n'
-        self.assertEqual(expected_text, feedback_correct(self.guess, quiz, self.nl))
+        self.assertEqual(expected_text, feedback_correct(self.guess, quiz, self.nl_fi))
 
     def test_show_alternative_answers(self):
         """Test that alternative answers are shown."""
         concept = self.create_concept("hi", dict(nl="hoi", fi=["terve", "hei", "hei hei"]))
-        quiz = create_quizzes(self.nl, self.fi, concept).by_quiz_type("read").pop()
+        quiz = create_quizzes(self.nl_fi, concept).by_quiz_type("read").pop()
         other_answers = [f'"{linkify(answer)}"' for answer in quiz.other_answers(self.guess)]
         expected_text = f'{CORRECT}[{SECONDARY}]Other correct answers are {", ".join(other_answers)}.[/{SECONDARY}]\n'
-        self.assertEqual(expected_text, feedback_correct(self.guess, quiz, self.nl))
+        self.assertEqual(expected_text, feedback_correct(self.guess, quiz, self.nl_fi))
 
     def test_show_feedback_on_incorrect_guess(self):
         """Test that the correct feedback is given when the user guesses incorrectly."""
         concept = self.create_concept("hi", dict(nl="hoi", fi="terve"))
-        quiz = create_quizzes(self.fi, self.nl, concept).by_quiz_type("dictate").pop()
+        quiz = create_quizzes(self.fi_nl, concept).by_quiz_type("dictate").pop()
         expected_text = (
             f'{INCORRECT}The correct answer is "[{INSERTED}]{linkify("terve")}[/{INSERTED}]".\n'
             f'[{SECONDARY}]Meaning "{linkify("hoi")}".[/{SECONDARY}]\n'
@@ -73,7 +74,7 @@ class FeedbackTestCase(ToistoTestCase):
     def test_show_alternative_answers_on_incorrect_guess(self):
         """Test that alternative answers are also given when the user guesses incorrectly."""
         concept = self.create_concept("hi", dict(nl="hoi", fi=["terve", "hei"]))
-        quiz = create_quizzes(self.nl, self.fi, concept).by_quiz_type("read").pop()
+        quiz = create_quizzes(self.nl_fi, concept).by_quiz_type("read").pop()
         expected_text = (
             f'{INCORRECT}The correct answer is "[{INSERTED}]{linkify("terve")}[/{INSERTED}]".\n'
             f'[{SECONDARY}]Another correct answer is "{linkify("hei")}".[/{SECONDARY}]\n'
@@ -83,21 +84,21 @@ class FeedbackTestCase(ToistoTestCase):
     def test_do_not_show_generated_alternative_answers_on_incorrect_guess(self):
         """Test that generated alternative answers are not shown when the user guesses incorrectly."""
         concept = self.create_concept("house", dict(nl="het huis", fi=["talo"]))
-        quiz = create_quizzes(self.fi, self.nl, concept).by_quiz_type("read").pop()
+        quiz = create_quizzes(self.fi_nl, concept).by_quiz_type("read").pop()
         expected_text = f'{INCORRECT}The correct answer is "[{INSERTED}]{linkify("het huis")}[/{INSERTED}]".\n'
         self.assertEqual(expected_text, feedback_incorrect(Label(self.nl, "incorrect"), quiz))
 
     def test_do_not_show_generated_alternative_answers_on_question_mark(self):
         """Test that generated alternative answers are not shown when the user enters a question mark."""
         concept = self.create_concept("house", dict(nl="het huis", fi=["talo"]))
-        quiz = create_quizzes(self.fi, self.nl, concept).by_quiz_type("read").pop()
+        quiz = create_quizzes(self.fi_nl, concept).by_quiz_type("read").pop()
         expected_text = f'The correct answer is "{linkify("het huis")}".\n'
         self.assertEqual(expected_text, feedback_incorrect(Label(self.nl, "?"), quiz))
 
     def test_show_feedback_on_question_mark(self):
         """Test that the correct feedback is given when the user doesn't know the answer."""
         concept = self.create_concept("hi", dict(nl="hoi", fi="terve"))
-        quiz = create_quizzes(self.fi, self.nl, concept).by_quiz_type("dictate").pop()
+        quiz = create_quizzes(self.fi_nl, concept).by_quiz_type("dictate").pop()
         expected_text = (
             f'The correct answer is "{linkify("terve")}".\n[{SECONDARY}]Meaning "{linkify("hoi")}".[/{SECONDARY}]\n'
         )
@@ -106,7 +107,7 @@ class FeedbackTestCase(ToistoTestCase):
     def test_show_feedback_on_question_mark_with_multiple_answers(self):
         """Test that the correct feedback is given when the user doesn't know the answer."""
         concept = self.create_concept("hi", dict(nl="hoi", fi=["terve", "hei"]))
-        quiz = create_quizzes(self.nl, self.fi, concept).by_quiz_type("read").pop()
+        quiz = create_quizzes(self.nl_fi, concept).by_quiz_type("read").pop()
         expected_text = (
             'The correct answers are "[link=https://en.wiktionary.org/wiki/terve]terve[/link]", '
             '"[link=https://en.wiktionary.org/wiki/hei]hei[/link]".\n'
@@ -116,7 +117,7 @@ class FeedbackTestCase(ToistoTestCase):
     def test_instruction(self):
         """Test that the quiz instruction is correctly formatted."""
         concept = self.create_concept("hi", dict(nl="hoi", fi="terve"))
-        quiz = create_quizzes(self.fi, self.nl, concept).by_quiz_type("write").pop()
+        quiz = create_quizzes(self.fi_nl, concept).by_quiz_type("write").pop()
         self.assertEqual(f"[{QUIZ}]Translate into Finnish:[/{QUIZ}]", instruction(quiz))
 
     def test_instruction_multiple_quiz_types(self):
@@ -125,7 +126,7 @@ class FeedbackTestCase(ToistoTestCase):
             "eat",
             {"first person": dict(nl="ik eet"), "third person": dict(female=dict(nl="zij eet"))},
         )
-        quizzes = create_quizzes(self.nl, self.nl, concept)
+        quizzes = create_quizzes(self.nl_fi, concept)
         quiz = quizzes.by_quiz_type("give third person").by_quiz_type("feminize").pop()
         expected_text = f"[{QUIZ}]Give the [underline]third person female[/underline] in Dutch:[/{QUIZ}]"
         self.assertEqual(expected_text, instruction(quiz))
@@ -133,25 +134,25 @@ class FeedbackTestCase(ToistoTestCase):
     def test_post_quiz_note(self):
         """Test that the post quiz note is formatted correctly."""
         concept = self.create_concept("hi", dict(nl="hoi;;Hoi is an informal greeting"))
-        quiz = create_quizzes(self.nl, self.nl, concept).by_quiz_type("dictate").pop()
+        quiz = create_quizzes(self.nl_fi, concept).by_quiz_type("dictate").pop()
         self.assertEqual(
             f"[{SECONDARY}]Note: Hoi is an informal greeting.[/{SECONDARY}]",
-            feedback_correct(Label(self.nl, "hoi"), quiz, self.nl).split("\n")[-2],
+            feedback_correct(Label(self.nl, "hoi"), quiz, self.nl_fi).split("\n")[-2],
         )
 
     def test_multiple_post_quiz_notes(self):
         """Test that multiple post quiz notes are formatted correctly."""
         concept = self.create_concept("hi", dict(fi="moi;;Moi is an informal greeting;'Moi moi' means goodbye"))
-        quiz = create_quizzes(self.fi, self.fi, concept).by_quiz_type("dictate").pop()
+        quiz = create_quizzes(self.fi_nl, concept).by_quiz_type("dictate").pop()
         self.assertIn(
             f"[{SECONDARY}]Notes:\n- Moi is an informal greeting.\n- 'Moi moi' means goodbye.[/{SECONDARY}]\n",
-            feedback_correct(Label(self.fi, "moi"), quiz, self.fi),
+            feedback_correct(Label(self.fi, "moi"), quiz, self.fi_nl),
         )
 
     def test_post_quiz_note_on_incorrect_answer(self):
         """Test that the post quiz note is formatted correctly."""
         concept = self.create_concept("hi", dict(fi="moi;;Moi is an informal greeting"))
-        quiz = create_quizzes(self.fi, self.fi, concept).by_quiz_type("dictate").pop()
+        quiz = create_quizzes(self.fi_nl, concept).by_quiz_type("dictate").pop()
         self.assertEqual(
             f"[{SECONDARY}]Note: Moi is an informal greeting.[/{SECONDARY}]",
             feedback_incorrect(Label(self.fi, "toi"), quiz).split("\n")[-2],
@@ -160,7 +161,7 @@ class FeedbackTestCase(ToistoTestCase):
     def test_post_quiz_note_on_skip_to_answer(self):
         """Test that the post quiz note is formatted correctly."""
         concept = self.create_concept("hi", dict(fi="moi;;Moi is an informal greeting"))
-        quiz = create_quizzes(self.fi, self.fi, concept).by_quiz_type("dictate").pop()
+        quiz = create_quizzes(self.fi_nl, concept).by_quiz_type("dictate").pop()
         self.assertEqual(
             f"[{SECONDARY}]Note: Moi is an informal greeting.[/{SECONDARY}]",
             feedback_incorrect(Label(self.fi, "?"), quiz).split("\n")[-2],
@@ -169,18 +170,24 @@ class FeedbackTestCase(ToistoTestCase):
     def test_post_quiz_example_with_spelling_alternatives(self):
         """Test that the post quiz example is formatted correctly when the example has spelling alternatives."""
         hi = self.create_concept("hi", dict(nl="hoi", fi="terve", example="hi alice"))
-        self.create_concept("hi alice", dict(fi="Moi Alice!|Hei Alice!"))
-        quiz = create_quizzes(self.fi, self.fi, hi).by_quiz_type("read").pop()
-        feedback_text = feedback_correct(self.guess, quiz, self.fi)
-        self.assertEqual(CORRECT + f"[{SECONDARY}]Example: Moi Alice![/{SECONDARY}]\n", feedback_text)
+        self.create_concept("hi alice", dict(fi="Moi Alice!|Hei Alice!", nl="Hoi Alice!"))
+        quiz = create_quizzes(self.fi_nl, hi).by_quiz_type("read").pop()
+        feedback_text = feedback_correct(Label(self.nl, "hoi"), quiz, self.fi_nl)
+        self.assertEqual(
+            CORRECT + f'[{SECONDARY}]Example: "Moi Alice!" meaning "Hoi Alice!"[/{SECONDARY}]\n',
+            feedback_text,
+        )
 
     def test_post_quiz_example_with_write_quiz(self):
         """Test that the post quiz example is in the right language when the quiz type is write."""
         hi = self.create_concept("hi", dict(nl="hoi", fi="terve", example="hi alice"))
         self.create_concept("hi alice", dict(fi="Terve Alice!", nl="Hoi Alice!"))
-        quiz = create_quizzes(self.fi, self.nl, hi).by_quiz_type("write").pop()
-        feedback_text = feedback_correct(self.guess, quiz, self.fi)
-        self.assertEqual(CORRECT + f"[{SECONDARY}]Example: Terve Alice![/{SECONDARY}]\n", feedback_text)
+        quiz = create_quizzes(self.fi_nl, hi).by_quiz_type("write").pop()
+        feedback_text = feedback_correct(self.guess, quiz, self.fi_nl)
+        self.assertEqual(
+            CORRECT + f'[{SECONDARY}]Example: "Terve Alice!" meaning "Hoi Alice!"[/{SECONDARY}]\n',
+            feedback_text,
+        )
 
 
 class LinkifyTest(TestCase):